pyproject.toml
[project]
name = "ai_assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuval"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "google-generativeai (>=0.5,<0.6)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"

